requirements-DESKTOP-C4FNGLN.txt
discord.py
python-dotenv
playwright
requests
matplotlib
beautifulsoup4
